+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.5.5] - 2024-07-10
 ### Added
 - Button to duplicate a page in a collection.
 - New option `nameField` for collections to use the value of a field as the document name [#19].
@@ -314,7 +314,7 @@ First version
 [#19]: https://github.com/lumeland/cms/issues/19
 [#20]: https://github.com/lumeland/cms/issues/20
 
-[Unreleased]: https://github.com/lumeland/cms/compare/v0.5.4...HEAD
+[0.5.5]: https://github.com/lumeland/cms/compare/v0.5.4...v0.5.5
 [0.5.4]: https://github.com/lumeland/cms/compare/v0.5.3...v0.5.4
 [0.5.3]: https://github.com/lumeland/cms/compare/v0.5.2...v0.5.3
 [0.5.2]: https://github.com/lumeland/cms/compare/v0.5.1...v0.5.2